# Conflicts:
#	hobbyhobby-user/hobbyhobby-storage/build/classes/java/main/org/v1/UserJpaEntity$UserJpaEntityBuilder.class
#	hobbyhobby-user/hobbyhobby-storage/build/classes/java/main/org/v1/UserJpaEntity.class
#	hobbyhobby-user/hobbyhobby-storage/build/libs/hobbyhobby-storage-0.0.1-SNAPSHOT-plain.jar
#	hobbyhobby-user/hobbyhobby-storage/build/tmp/compileJava/previous-compilation-data.bin

+++ hobbyhobby-community/hobbyhobby-domain/src/main/java/org/v1/service/UserCommunityService.java
@@ -10,10 +10,10 @@ import org.v1.implementation.community.UserCommunityRemover;
 public class UserCommunityService {
     private final UserCommunityRemover userCommunityRemover;
     private final UserCommunityAppender userCommunityAppender;
-    public void removeUserFromCommunity(Long userId, Long communityId) {
+    public void removeUserFromCommunity(Long userId, Integer communityId) {
         userCommunityRemover.removeUserFromCommunity(userId, communityId);
     }
-    public void createUserToCommunity(Long userId, Long communityId) {
+    public void createUserToCommunity(Long userId, Integer communityId) {
         userCommunityAppender.appendUserToCommunity(userId, communityId);
     }
 }
